ファイルドロップ対応
read時にドロップしたファイルのデータを表示できるように対応した
クラス構成を大幅に改修
