+++ device_control_server/ic_card/ic_card_001.yml
@@ -1,0 +1,311 @@
+# coding: utf-8
+content-type: pasori_history
+content-version: 0.1
+
+contents:
+  description: "標準的なカード 000"
+  read_status: 0 # 0:読み込み成功, 1:読み込み失敗
+  pasori_data:
+    idm: "0000000000000000"
+    pmm: "1234567890123456"
+    history:
+      - ctype: 1
+        ctype_name: "改札機"
+        proc: 2
+        proc_name: "運賃支払"
+        date: 3
+        date_string: "14/04/01"
+        time: 4
+        time_string: "12:34"
+        balance: 1000
+        region: 5
+        seq: 6
+        in_line: 100
+        in_sta: 101
+        out_line: 200
+        out_sta: 201
+      - ctype: 1
+        ctype_name: "改札機"
+        proc: 2
+        proc_name: "運賃支払"
+        date: 3
+        date_string: "14/04/01"
+        time: 4
+        time_string: "12:34"
+        balance: 1000
+        region: 5
+        seq: 6
+        in_line: 100
+        in_sta: 101
+        out_line: 200
+        out_sta: 201
+      - ctype: 1
+        ctype_name: "改札機"
+        proc: 2
+        proc_name: "運賃支払"
+        date: 3
+        date_string: "14/04/01"
+        time: 4
+        time_string: "12:34"
+        balance: 1000
+        region: 5
+        seq: 6
+        in_line: 100
+        in_sta: 101
+        out_line: 200
+        out_sta: 201
+      - ctype: 1
+        ctype_name: "改札機"
+        proc: 2
+        proc_name: "運賃支払"
+        date: 3
+        date_string: "14/04/01"
+        time: 4
+        time_string: "12:34"
+        balance: 1000
+        region: 5
+        seq: 6
+        in_line: 100
+        in_sta: 101
+        out_line: 200
+        out_sta: 201
+      - ctype: 1
+        ctype_name: "改札機"
+        proc: 2
+        proc_name: "運賃支払"
+        date: 3
+        date_string: "14/04/01"
+        time: 4
+        time_string: "12:34"
+        balance: 1000
+        region: 5
+        seq: 6
+        in_line: 100
+        in_sta: 101
+        out_line: 200
+        out_sta: 201
+      - ctype: 1
+        ctype_name: "改札機"
+        proc: 2
+        proc_name: "運賃支払"
+        date: 3
+        date_string: "14/04/01"
+        time: 4
+        time_string: "12:34"
+        balance: 1000
+        region: 5
+        seq: 6
+        in_line: 100
+        in_sta: 101
+        out_line: 200
+        out_sta: 201
+      - ctype: 1
+        ctype_name: "改札機"
+        proc: 2
+        proc_name: "運賃支払"
+        date: 3
+        date_string: "14/04/01"
+        time: 4
+        time_string: "12:34"
+        balance: 1000
+        region: 5
+        seq: 6
+        in_line: 100
+        in_sta: 101
+        out_line: 200
+        out_sta: 201
+      - ctype: 1
+        ctype_name: "改札機"
+        proc: 2
+        proc_name: "運賃支払"
+        date: 3
+        date_string: "14/04/01"
+        time: 4
+        time_string: "12:34"
+        balance: 1000
+        region: 5
+        seq: 6
+        in_line: 100
+        in_sta: 101
+        out_line: 200
+        out_sta: 201
+      - ctype: 1
+        ctype_name: "改札機"
+        proc: 2
+        proc_name: "運賃支払"
+        date: 3
+        date_string: "14/04/01"
+        time: 4
+        time_string: "12:34"
+        balance: 1000
+        region: 5
+        seq: 6
+        in_line: 100
+        in_sta: 101
+        out_line: 200
+        out_sta: 201
+      - ctype: 1
+        ctype_name: "改札機"
+        proc: 2
+        proc_name: "運賃支払"
+        date: 3
+        date_string: "14/04/01"
+        time: 4
+        time_string: "12:34"
+        balance: 1000
+        region: 5
+        seq: 6
+        in_line: 100
+        in_sta: 101
+        out_line: 200
+        out_sta: 201
+      - ctype: 1
+        ctype_name: "改札機"
+        proc: 2
+        proc_name: "運賃支払"
+        date: 3
+        date_string: "14/04/01"
+        time: 4
+        time_string: "12:34"
+        balance: 1000
+        region: 5
+        seq: 6
+        in_line: 100
+        in_sta: 101
+        out_line: 200
+        out_sta: 201
+      - ctype: 1
+        ctype_name: "改札機"
+        proc: 2
+        proc_name: "運賃支払"
+        date: 3
+        date_string: "14/04/01"
+        time: 4
+        time_string: "12:34"
+        balance: 1000
+        region: 5
+        seq: 6
+        in_line: 100
+        in_sta: 101
+        out_line: 200
+        out_sta: 201
+      - ctype: 1
+        ctype_name: "改札機"
+        proc: 2
+        proc_name: "運賃支払"
+        date: 3
+        date_string: "14/04/01"
+        time: 4
+        time_string: "12:34"
+        balance: 1000
+        region: 5
+        seq: 6
+        in_line: 100
+        in_sta: 101
+        out_line: 200
+        out_sta: 201
+      - ctype: 1
+        ctype_name: "改札機"
+        proc: 2
+        proc_name: "運賃支払"
+        date: 3
+        date_string: "14/04/01"
+        time: 4
+        time_string: "12:34"
+        balance: 1000
+        region: 5
+        seq: 6
+        in_line: 100
+        in_sta: 101
+        out_line: 200
+        out_sta: 201
+      - ctype: 1
+        ctype_name: "改札機"
+        proc: 2
+        proc_name: "運賃支払"
+        date: 3
+        date_string: "14/04/01"
+        time: 4
+        time_string: "12:34"
+        balance: 1000
+        region: 5
+        seq: 6
+        in_line: 100
+        in_sta: 101
+        out_line: 200
+        out_sta: 201
+      - ctype: 1
+        ctype_name: "改札機"
+        proc: 2
+        proc_name: "運賃支払"
+        date: 3
+        date_string: "14/04/01"
+        time: 4
+        time_string: "12:34"
+        balance: 1000
+        region: 5
+        seq: 6
+        in_line: 100
+        in_sta: 101
+        out_line: 200
+        out_sta: 201
+      - ctype: 1
+        ctype_name: "改札機"
+        proc: 2
+        proc_name: "運賃支払"
+        date: 3
+        date_string: "14/04/01"
+        time: 4
+        time_string: "12:34"
+        balance: 1000
+        region: 5
+        seq: 6
+        in_line: 100
+        in_sta: 101
+        out_line: 200
+        out_sta: 201
+      - ctype: 1
+        ctype_name: "改札機"
+        proc: 2
+        proc_name: "運賃支払"
+        date: 3
+        date_string: "14/04/01"
+        time: 4
+        time_string: "12:34"
+        balance: 1000
+        region: 5
+        seq: 6
+        in_line: 100
+        in_sta: 101
+        out_line: 200
+        out_sta: 201
+      - ctype: 1
+        ctype_name: "改札機"
+        proc: 2
+        proc_name: "運賃支払"
+        date: 3
+        date_string: "14/04/01"
+        time: 4
+        time_string: "12:34"
+        balance: 1000
+        region: 5
+        seq: 6
+        in_line: 100
+        in_sta: 101
+        out_line: 200
+        out_sta: 201
+      - ctype: 1
+        ctype_name: "改札機"
+        proc: 2
+        proc_name: "運賃支払"
+        date: 3
+        date_string: "14/04/01"
+        time: 4
+        time_string: "12:34"
+        balance: 1000
+        region: 5
+        seq: 6
+        in_line: 100
+        in_sta: 101
+        out_line: 200
+        out_sta: 201
