+++ device_control_server/ic_card/ic_card_002.yml
@@ -1,25 +1,13 @@
 # coding: utf-8
-content-type: ic_card
+content-type: pasori_history
 content-version: 0.1
 
 contents:
   description: "標準的なカード 000"
   read_status: 0 # 0:読み込み成功, 1:読み込み失敗
-  contents:
-    idi: "JP0000"
-    name: "テスト たろう"
-    birthday: "2000/1/1 10:00"
-    money_value: 1000
-    ic_status: 0
-    start_station: "001-001"
-    end_station: "001-002"
-    start_date: "2015/4/1"
-    end_date: "20015/8/31"
-    write_counter: 0
-  pasori_contents:
-    status: 0
-    idm: [00,00,00,00,00,00,00,00]
-    pmm: [12,34,56,78,90,12,34,56]
+  pasori_data:
+    idm: "0000000000000000"
+    pmm: "1234567890123456"
     history:
       - ctype: 1
         ctype_name: "改札機"
