+++ device_control_server/parameter.yml
@@ -4,5 +4,5 @@ content-version: 0.1
 
 contents:
   description: "qt server パラメータ一式"
-  ip: '192.168.10.58'
+  ip: '0.0.0.0'
   port: '3001'
