+++ public/index.html
@@ -16,7 +16,7 @@
     <div class="form-inline">
       <div class="form-group">
         <label>PaSoRi端末接続先：</label>
-        <input type="text" class="form-control pasori-address" id="exampleInputName2" placeholder="172.0.0.1:3001" value="172.17.10.58:3001">
+        <input type="text" class="form-control pasori-address" id="exampleInputName2" placeholder="172.0.0.1:3001" value="localhost:3001">
       </div>
     </div>
